+++ relay-general/src/protocol/contexts/mod.rs
@@ -91,13 +91,13 @@ impl Context {
 /// Trait to get the Context both from the user agent string and also the new client hints.
 /// With an automatically derived function which tries to first get the context from client hints,
 /// if that fails it tries for the user agent string.
-pub trait ContextFromUserAgentInfo: Sized {
-    fn from_client_hints(raw_contexts: &RawUserAgentInfo) -> Option<Self>;
+pub trait FromUserAgentInfo: Sized {
+    fn from_client_hints(raw_info: &RawUserAgentInfo) -> Option<Self>;
     fn from_user_agent(user_agent: &str) -> Option<Self>;
 
-    fn from_hints_or_ua(raw_contexts: &RawUserAgentInfo) -> Option<Self> {
-        Self::from_client_hints(raw_contexts)
-            .or_else(|| raw_contexts.user_agent.and_then(Self::from_user_agent))
+    fn from_hints_or_ua(raw_info: &RawUserAgentInfo) -> Option<Self> {
+        Self::from_client_hints(raw_info)
+            .or_else(|| raw_info.user_agent.and_then(Self::from_user_agent))
     }
 }