+++ relay-general/src/protocol/contexts/os.rs
@@ -1,4 +1,4 @@
-use crate::protocol::{ContextFromUserAgentInfo, LenientString};
+use crate::protocol::{FromUserAgentInfo, LenientString};
 use crate::store::user_agent::{get_version, is_known};
 use crate::types::{Annotated, Object, Value};
 use crate::user_agent::{parse_os, RawUserAgentInfo};
@@ -49,10 +49,10 @@ impl OsContext {
     }
 }
 
-impl ContextFromUserAgentInfo for OsContext {
+impl FromUserAgentInfo for OsContext {
     fn from_client_hints(contexts: &RawUserAgentInfo) -> Option<Self> {
-        let platform = contexts.sec_ch_ua_platform?;
-        let version = contexts.sec_ch_ua_platform_version?;
+        let platform = contexts.client_hints.sec_ch_ua_platform?;
+        let version = contexts.client_hints.sec_ch_ua_platform_version?;
 
         Some(Self {
             name: Annotated::new(platform.to_owned()),