+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 **Features**:
 
 - Add error and sample rate fields to the replay event parser. ([#1745](https://github.com/getsentry/relay/pull/1745))
+- Add support for client hints. ([#1752](https://github.com/getsentry/relay/pull/1752))
 
 **Internal**: